+++ dmr/validation/payload.py
@@ -10,6 +10,7 @@ from dmr.metadata import ResponseSpec
 from dmr.parsers import Parser
 from dmr.renderers import Renderer
 from dmr.settings import HttpSpec
+from dmr.types import Empty, EmptyObj
 
 if TYPE_CHECKING:
     from dmr.openapi.objects import (
@@ -49,6 +50,7 @@ class _BasePayload:
     validate_negotiation: bool | None = None
     auth: Sequence['SyncAuth'] | Sequence['AsyncAuth'] | None = ()
     throttling: Sequence['SyncThrottle'] | Sequence['AsyncThrottle'] | None = ()
+    throttling_allow_unsafe_cache: bool | Empty | None = EmptyObj
 
 
 @dataclasses.dataclass(slots=True, frozen=True, kw_only=True)