+++ tests/test_unit/test_streaming/test_sse/test_class_sse.py
@@ -17,7 +17,7 @@ _EventsType: TypeAlias = SSEvent[dict[str, str] | bytes]
 
 async def _valid_events() -> AsyncIterator[_EventsType]:
     yield SSEvent({})
-    yield SSEvent(b'multiline\nbyte\nstring', serialize=False)  # pyrefly: ignore[no-matching-overload]
+    yield SSEvent(b'multiline\nbyte\nstring', serialize=False)
 
 
 class _ClassBasedSSE(SSEController[PydanticSerializer]):